refactor(select): migrate selection UI from forge_main to forge_select crate
