+++ crates/forge_select/Cargo.toml
@@ -6,8 +6,11 @@ rust-version.workspace = true
 
 [dependencies]
 anyhow.workspace = true
+bstr.workspace = true
 colored.workspace = true
 console.workspace = true
+crossterm = "0.29.0"
+nucleo.workspace = true
 nucleo-picker.workspace = true
 rustyline.workspace = true
 tracing.workspace = true
