+++ crates/forge_select/src/lib.rs
@@ -1,10 +1,14 @@
 mod confirm;
 mod input;
 mod multi;
+mod preview;
 mod select;
 mod widget;
 
 pub use input::InputBuilder;
 pub use multi::MultiSelectBuilder;
+pub use preview::{
+    PreviewLayout, PreviewPlacement, SelectMode, SelectRow, SelectUiOptions, run_select_ui,
+};
 pub use select::SelectBuilder;
 pub use widget::ForgeWidget;
